Switch to the 3-argument version of timespecadd
nstat defines a private 2-argument timespecadd macro that's a copy of
the FreeBSD kernel's.  But NetBSD, OpenBSD, and FreeDesktop.org's libbsd
all define a public 3-argument timespecadd macro instead.  An upcoming
change to FreeBSD will replace the 2-argument version with the more
common 3-argument version and make it public.

This change brings nstat onto the 3-argument bandwagon and ensures that
it will continue to build on old versions of FreeBSD too.